+++ legacy/java/allure-notifications-api/src/main/java/guru/qa/allure/notifications/config/enums/Language.java
@@ -1,5 +1,5 @@
 package guru.qa.allure.notifications.config.enums;
 
 public enum Language {
-    en, fr, ru, ua, cn, cnt, by
+    en, de, fr, ru, by, ua, cn, cnt
 }