+++ packages/config/src/schema.ts
@@ -5,7 +5,7 @@
 
 import { z } from "zod";
 
-const LanguageSchema = z.enum(["en", "fr", "ru", "ua", "cn", "cnt", "by"]);
+const LanguageSchema = z.enum(["en", "de", "fr", "ru", "by", "ua", "cn", "cnt"]);
 
 const LinksSchema = z
   .object({